requirements.txt
# === Core Framework ===
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
email-validator>=2.0.0

# === Database ===
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# === Caching & Sessions ===
redis>=5.0.0

# === Security ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0

# === AI & LLM ===
anthropic>=0.18.0
openai>=1.12.0

# === Vector Database ===
pinecone-client>=3.0.0

# === Blockchain ===
web3>=6.14.0
base58>=2.1.1

# === MPC Libraries ===
pynacl>=1.5.0
ecdsa>=0.18.0

# === HTTP Clients ===
httpx>=0.26.0
aiohttp>=3.9.0

# === Utilities ===
python-dotenv>=1.0.0
structlog>=24.1.0
orjson>=3.9.0
